+++ ios/RCTBackgroundGeolocation/LocationContract.h
@@ -22,6 +22,7 @@
 #define LC_COLUMN_NAME_PROVIDER            "provider"
 #define LC_COLUMN_NAME_LOCATION_PROVIDER   "service_provider"
 #define LC_COLUMN_NAME_VALID               "valid"
+#define LC_COLUMN_NAME_RECORDED_AT         "recorded_at"
 
 @interface LocationContract : NSObject