+++ ios/RCTBackgroundGeolocation/RCTBackgroundGeolocation.h
@@ -7,14 +7,14 @@
 //
 
 #if __has_include(<React/RCTBridgeModule.h>)
-  #import <React/RCTBridgeModule.h>
+#import <React/RCTBridgeModule.h>
 #else
-  #import "RCTBridgeModule.h"
+#import "RCTBridgeModule.h"
 #endif
-#import "LocationManager.h"
+#import "BackgroundGeolocationFacade.h"
 
-@interface RCTBackgroundGeolocation : NSObject <RCTBridgeModule, LocationManagerDelegate>
+@interface RCTBackgroundGeolocation : NSObject <RCTBridgeModule, LocationDelegate>
 
-@property (nonatomic, strong) LocationManager* locationManager;
+@property (nonatomic, strong) BackgroundGeolocationFacade* facade;
 
 @end