+++ CHANGES.md
@@ -1,5 +1,26 @@
 ## Changelog
 
+### [0.4.0-rc.1] - 2017-11-10
+
+### Fixed
+- iOS fix crash when calling getConfig before configure
+
+#### Added
+- checkStatus if service is running
+- events [start, stop, authorization, background, foreground]
+- implement all methods for both platforms
+- new RAW_LOCATION_PROVIDER
+
+#### Changed
+
+- start and stop methods doesn't accept callback (use event listeners instead)
+- for background syncing syncUrl option is required
+- on Android DISTANCE_FILTER_PROVIDER now accept arbitrary values (before only 10, 100, 1000)
+- all plugin constants are in directly BackgroundGeolocation namespace. (check index.js)
+- plugin can be started without executing configure (stored settings or defaults will be used)
+- location property locationId renamed to just id
+- iOS pauseLocationUpdates now default to false (becuase iOS docs now states that you need to restart manually if you set it to true)
+
 ### [0.3.3] - 2017-11-01
 #### Fixed
 - Android location sync should also be completed on 201 status code (PR #71)